CICD_Databricks/requirements.txt
# =============================================================================
# Python Dependencies
# =============================================================================
# Requirements for local development and CI/CD pipeline execution.
# =============================================================================

# -----------------------------------------------------------------------------
# Core Dependencies
# -----------------------------------------------------------------------------
pyspark>=3.5.0
delta-spark>=3.0.0
databricks-sdk>=0.20.0

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-html>=4.1.0
pytest-xdist>=3.5.0  # Parallel test execution

# -----------------------------------------------------------------------------
# Code Quality
# -----------------------------------------------------------------------------
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
pyyaml>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# -----------------------------------------------------------------------------
# Data Quality (Optional - for local DQx development)
# -----------------------------------------------------------------------------
great_expectations>=0.18.0
